Allow deployment from a renamed deploy directory.
The source directory of the deploy/deploy.sh script is captured and passed to the deploy.sh and copy_files.sh scripts in the cfs folder.

This is useful if you want to have different configurations stored in the same repository, as it is the case with jd sport, for example.

Change-Id: I129450627d4cde2922d5c022444a1c331c0fa61c